move meme functions to a meme file
