+++ config/config.go
@@ -1,7 +1,12 @@
 package config
 
 import (
+	"encoding/json"
+	"fmt"
+	"io"
+	"log"
 	"math/rand"
+	"net/http"
 	"os"
 	"time"
 )
@@ -23,7 +28,11 @@ func InitConfig() {
 	RelayUrl = os.Getenv("RELAY_URL")
 	MemeUrl = os.Getenv("MEME_URL")
 	RelayAuthKey = os.Getenv("RELAY_AUTH_KEY")
-	RelayNodeKey = os.Getenv("RELAY_NODE_KEY")
+
+	// only make this call if there is a Relay auth key
+	if RelayAuthKey != "" {
+		RelayNodeKey = GetNodePubKey()
+	}
 
 	if Host == "" {
 		Host = "https://people.sphinx.chat"
@@ -53,4 +62,68 @@ func GenerateRandomString() string {
 	}
 
 	return string(b)
+}
+
+type PropertyMap map[string]interface{}
+
+type Feature map[string]PropertyMap
+
+type NodeGetInfoResponse struct {
+	Uris                   []string      `json:"uris"`
+	Chains                 []PropertyMap `json:"chains"`
+	Features               Feature       `json:"features"`
+	IdentityPubkey         string        `json:"identity_pubkey"`
+	Alias                  string        `json:"alias"`
+	NumPendingChannels     uint          `json:"num_pending_channels"`
+	NumActiveChannels      uint          `json:"num_active_channels"`
+	NumInactiveChannels    uint          `json:"num_inactive_channels"`
+	NumPeers               uint          `json:"num_peers"`
+	BlockHeight            uint          `json:"block_height"`
+	BlockHash              string        `json:"block_hash"`
+	SyncedToChain          bool          `json:"synced_to_chain"`
+	Testnet                bool          `json:"testnet"`
+	BestHeaderTimestamp    string        `json:"best_header_timestamp"`
+	Version                string        `json:"version"`
+	Color                  string        `json:"color"`
+	SyncedToGraph          bool          `json:"synced_to_graph"`
+	CommitHash             string        `json:"commit_hash"`
+	RequireHtlcInterceptor bool          `json:"require_htlc_interceptor"`
+}
+
+type NodeGetInfo struct {
+	Success  bool                `json:"success"`
+	Response NodeGetInfoResponse `json:"response"`
+}
+
+func GetNodePubKey() string {
+	var pubkey string
+	url := fmt.Sprintf("%s/getinfo", RelayUrl)
+
+	client := &http.Client{}
+	req, err := http.NewRequest(http.MethodGet, url, nil)
+
+	req.Header.Set("x-user-token", RelayAuthKey)
+	req.Header.Set("Content-Type", "application/json")
+	res, _ := client.Do(req)
+
+	if err != nil {
+		log.Printf("Request Failed: %s", err)
+	}
+
+	defer res.Body.Close()
+
+	body, err := io.ReadAll(res.Body)
+
+	nodeInfo := NodeGetInfo{}
+
+	// Unmarshal result
+	err = json.Unmarshal(body, &nodeInfo)
+
+	if err != nil {
+		log.Printf("Reading Relay Node Info body failed: %s", err)
+	}
+
+	pubkey = nodeInfo.Response.IdentityPubkey
+
+	return pubkey
 }
